fixes nasty bug in IncludeLoader: return absolute resource uris
the bootstrap template is itself loading other templates to include them. Due
to a bug in include loader, a uri to a non-existing resources can be returned.
This leads to a ResourceNotFoundExc. in the DefaultLayoutStrategy. This in
turn invalidates the layout (removes it from the context) in the middle of
rendering. The oucome was a messed up page (the body content was
rendered inside the nav bar div).